api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pyodbc==5.1.0
python-dotenv==1.0.1